feat: MySQL support, public landing page, improved fetch script
Database:
- Switch from SQLite to MySQL 8.0+ for production readiness
- Change timestamp columns to BigInteger for millisecond timestamps
- Add split database config (DB_* for dev, PROD_DB_* for prod)
- Auto-detect production environment (FLASK_ENV, PaaS indicators)

Routing:
- Make landing page public at / (no login required)
- Move dashboard to /dashboard prefix
- Logged-in users redirected from / to /dashboard

Fetch Script (fetch_historical.py):
- Use fetch_start_date setting from Admin > Symbols (default: 2024-01-01)
- Show symbol names clearly in output
- Serialize parallel output with asyncio.Lock to prevent interleaving
- --days flag now overrides the database setting
- Single "Logging system active" message (reuse app context)

Documentation:
- Update README with MySQL setup, Makefile commands
- Update admin documentation with MySQL references
- Add database configuration section with PROD_DB_* vars

+++ app/models/system.py
@@ -76,7 +76,7 @@ class StatsCache(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     key = db.Column(db.String(50), unique=True, nullable=False)  # 'global' or symbol name
     data = db.Column(db.Text, nullable=False)  # JSON blob
-    computed_at = db.Column(db.Integer, nullable=False)  # Timestamp in ms
+    computed_at = db.Column(db.BigInteger, nullable=False)  # Timestamp in ms
 
     def __repr__(self):
         return f'<StatsCache {self.key}>'
